pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
dnspython==2.8.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gcsfs==2026.1.0
google-api-core==2.28.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-cloud==0.34.0
google-cloud-bigquery==3.40.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-cloud-storage-control==1.8.0
google-crc32c==1.8.0
google-resumable-media
googleapis-common-protos==1.72.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
grpcio-tools==1.76.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
neo4j==6.1.0
networkx==3.6.1
numpy==2.4.1
oauthlib==3.3.1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymilvus==2.6.6
pymilvus[milvus_lite]
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
yarl==1.22.0
